% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjm}
\alias{adjm}
\title{A fitted Bayesian multilevel compositional model using 'brmcoda'
An example of adjusted 'brmcoda' object}
\format{
A list with 2 elemets.
\describe{
\item{CompILR}{A 'compilr' object}
\item{Model}{Fitted brm multivariate multilevel
mmodel with compositional outcome}
}
}
\usage{
adjm
}
\description{
A fitted Bayesian multilevel compositional model using 'brmcoda'
An example of adjusted 'brmcoda' object
}
\keyword{datasets}
