% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjmvm}
\alias{adjmvm}
\title{A fitted Bayesian multivariate multilevel compositional model using 'mvmcoda'
An example of adjusted 'mvmcoda' object}
\format{
A list with 2 elemets.
\describe{
\item{CompILR}{A 'compilr' object}
\item{Model}{Fitted brm multivariate multilevel
mmodel with compositional outcome}
}
}
\usage{
adjmvm
}
\description{
A fitted Bayesian multivariate multilevel compositional model using 'mvmcoda'
An example of adjusted 'mvmcoda' object
}
\keyword{datasets}
