% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complr-methods.R
\name{as.complr}
\alias{as.complr}
\title{Coerce a list to a \code{complr} object}
\usage{
as.complr(x)
}
\arguments{
\item{x}{A list with elements \code{output}, \code{datain}, \code{dataout},
\code{transform}, and \code{idvar}.}
}
\value{
A \code{complr} object.
}
\description{
This is a constructor function for a \code{complr} object.
It checks that the input list has the necessary structure and
components to be considered a \code{complr} object,
and if so, it assigns the class "complr" to it.
This allows for method dispatch on \code{complr} objects.
}
