% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{as.data.frame.brmcoda}
\alias{as.data.frame.brmcoda}
\alias{as.matrix.brmcoda}
\alias{as.array.brmcoda}
\title{Extract Posterior Draws}
\usage{
\method{as.data.frame}{brmcoda}(x)

\method{as.matrix}{brmcoda}(x)

\method{as.array}{brmcoda}(x)
}
\arguments{
\item{x}{An object of class \code{brmcoda}.}
}
\description{
Extract posterior draws in conventional formats
as data.frames, matrices, or arrays.
}
