% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilr-methods.R
\name{as.data.frame.compilr}
\alias{as.data.frame.compilr}
\alias{as.matrix.compilr}
\title{Extract Compositional Data from \code{compilr} object.}
\usage{
\method{as.data.frame}{compilr}(
  x,
  class = c("composition", "logratio"),
  level = c("between", "within", "total"),
  digits = 3,
  ...
)

\method{as.matrix}{compilr}(
  x,
  class = c("composition", "logratio"),
  level = c("between", "within", "total"),
  digits = 3,
  ...
)
}
\arguments{
\item{x}{An object of class \code{compilr}.}

\item{class}{Optional. Can be \code{"composition"} and/or \code{"logratio"} to
specify the geometry of the composition.}

\item{level}{Optional. Can be \code{"between"}, \code{"within"}, and/or \code{total}
indicating the level of the geometry.}

\item{digits}{A integer value used for number formatting. Default is \code{3}.}
}
\description{
Extract amounts and compositions in conventional formats
as data.frames, matrices, or arrays.
}
