% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{as.diagnostics}
\alias{as.diagnostics}
\title{Coerce a list to a \code{diagnostics} object}
\usage{
as.diagnostics(x)
}
\arguments{
\item{x}{A list with elements \code{x}, \code{distance}, \code{cutoff},
\code{ev.perc}, \code{extremevalues}, and \code{levels}.}
}
\value{
A \code{diagnostics} object.
}
\description{
This is a constructor function for a \code{diagnostics} object.
It checks that the input list has the necessary structure and
components to be considered a \code{diagnostics} object,
and if so, it assigns the class "diagnostics" to it.
This allows for method dispatch on \code{diagnostics} objects.
}
