% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basesub.R
\name{basesub}
\alias{basesub}
\title{Base Pairwise Substitution}
\usage{
basesub(parts)
}
\arguments{
\item{parts}{A character vector specifying the names of compositional variables to be used.}
}
\value{
A data table of all possible pairwise substitution.
}
\description{
Make a data set of all possible pairwise substitution of a composition which can be used as
the base for substitution models.
}
\examples{
ps1 <- basesub(parts = c("TST", "WAKE", "MVPA", "LPA", "SB"))
print(ps1)

ps2 <- basesub(c("WAKE", "MVPA", "LPA", "SB"))
print(ps2)
}
