% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{bayes_factor.brmcoda}
\alias{bayes_factor.brmcoda}
\title{Bayes Factors from Marginal Likelihoods}
\usage{
\method{bayes_factor}{brmcoda}(x1, x2, ...)
}
\description{
Bayes Factors from Marginal Likelihoods
}
\seealso{
\code{\link[brms:bayes_factor.brmsfit]{bayes_factor.brmsfit}}
}
