% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsmodel.R
\name{brmcoda}
\alias{brmcoda}
\title{Computing brms model for compostion predicting an outcome}
\usage{
brmcoda(compilr, formula, data)
}
\arguments{
\item{compilr}{used in the model.}

\item{formula}{A symbolic description of the model to be fitted.}

\item{data}{containing data of all variables other than ILR coordinates used in the model.}
}
\value{
A list with two elements
\itemize{
\item{\code{CompIlr}}{ A object of class \code{compilr} used in the \code{brm} model. }
\item{\code{Results}}{ An object of class \code{brmsfit}, which contains the posterior draws along with many other useful information about the model.}
}
\description{
Computing brms model for compostion predicting an outcome
}
\examples{
data(mcompd)
data(sbp)

## compute compositions and ILR coordinates
compilrtest <- compilr(data = mcompd[, 1:6], sbp = sbp, idvar = "ID")

## function testing
brmcodatest <- brmcoda(data = mcompd, compilr = compilrtest, 
formula = STRESS ~ bilr1 + bilr2 + bilr3 + bilr4 + wilr1 + wilr2 + wilr3 + wilr4 + (1 | ID))

## clean-up
rm(mcompd, sbp, compilrtest, brmcodatest)
}
