% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsub.R, R/bsubmargins.R, R/wsub.R
\name{bsubmargin}
\alias{bsubmargin}
\title{Between-person Substitution Model (from sample's compositional mean)}
\usage{
bsubmargin(object, b, comp, at, names, substitute)

bsubmargin(object, b, comp, at, names, substitute)

bsubmargin(object, b, comp, at, names, substitute)
}
\arguments{
\item{object}{A fitted \code{brms} model object. Required.}

\item{b}{A compisition or dataset of composition at between-peson level. Required.}

\item{comp}{A data frame or data table indicating the possible substitution of variables.}

\item{at}{Default is 1:60L.}

\item{names}{A vector representing names of variables in the substitution data.}

\item{substitute}{A vector representing names of substituted compositional component for each row of substitution dataset.}
}
\value{




TODO
}
\description{
Estimate the difference in outcomes
when compositional variables are substituted for a specific time period.
at between-person level.

Estimate the difference in outcomes
when compositional variables are substituted for a specific time period
at between-person level.

Estimate the difference in outcomes
when compositional variables are substituted for a specific time period
at between-person level.
}
\examples{
## TODO
## TODO
## subd <- bsubmargin(object = m, b = b, comp = comp)
## TODO
##subd <- bsubmargin(object = m, b = b, comp = comp)
}
