% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complr.R
\name{complr}
\alias{complr}
\title{Indices from a (dataset of) Multilevel Composition(s)}
\usage{
complr(data, parts, transform = "ilr", sbp = NULL, total = 1, idvar = "ID")
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table}
containing data of all variables used in the analysis.
It must include a composition and a ID variable. Required.}

\item{parts}{A character vector specifying the names of compositional variables to be used.}

\item{transform}{A character value naming a log ratio transformation to be applied on compositional data.
Can be either \code{"ilr"} (isometric logratio), \code{"alr"} (additive logratio), or \code{"clr"} (centered logratio).
Default is \code{"ilr"}.}

\item{sbp}{A signary matrix indicating sequential binary partition. Required.}

\item{total}{A numeric value of the total amount to which the compositions should be closed.
Default is \code{1}.}

\item{idvar}{A character string specifying the name of the variable containing IDs.
Default is \code{"ID"}.}
}
\value{
A \code{\link{complr}} object with twelve elements.
\item{\code{BetweenComp}}{ A vector of class \code{acomp} representing one closed between-person composition
or a matrix of class \code{acomp} representing multiple closed between-person compositions each in one row.}
\item{\code{WithinComp}}{ A vector of class \code{acomp} representing one closed within-person composition
or a matrix of class \code{acomp} representing multiple closed within-person compositions each in one row.}
\item{\code{TotalComp}}{ A vector of class \code{acomp} representing one closed total composition
or a matrix of class \code{acomp} representing multiple closed total compositions each in one row.}
\item{\code{BetweenILR}}{ Isometric log ratio transform of between-person composition.}
\item{\code{WithinILR}}{ Isometric log ratio transform of within-person composition.}
\item{\code{TotalILR}}{ Isometric log ratio transform of total composition.}
\item{\code{data}}{ The user's dataset or imputed dataset if the input data contains zeros.}
\item{\code{psi}}{ A ILR matrix associated with user-defined partition structure.}
\item{\code{sbp}}{ The user-defined sequential binary partition matrix.}
\item{\code{parts}}{ Names of compositional variables.}
\item{\code{idvar}}{ Name of the variable containing IDs.}
\item{\code{total}}{ Total amount to which the compositions is closed.}
}
\description{
Compute sets of compositions and log ratio transformation for multilevel compositional data
}
\details{
The \emph{ilr}-transform maps the D-part compositional data from the simplex into non-overlapping
subgroups in the (D−1)-dimension Euclidean space isometrically by using an orthonormal basis,
thereby preserving the compo- sitional properties and yielding a full-rank covariance matrix.
\emph{ilr} transformation should be preferred.
However, the \emph{alr} and \emph{clr} are alternatives.
The \emph{alr}-transform maps a D-part composition
in the Aitchison-simplex non-isometrically to a
D − 1-dimension Euclidian vectors,
commonly treating the last part as the common denominator of the others.
\emph{alr} transformation does not rely on distance which breaks
the constraint of compositional data.
\emph{clr}-transform maps a D-part composition in the Aitchison-simplex
isometrically to a D-dimensional Euclidian vector subspace.
\emph{clr} transformation is not injetive,
resulting in singular covariance matrices.
}
\examples{
cilr <- complr(data = mcompd, sbp = sbp,
                parts = c("TST", "WAKE", "MVPA", "LPA", "SB"),
                idvar = "ID", total = 1440)
str(cilr)
}
