% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{diagnostic-quantities-brmcoda}
\alias{diagnostic-quantities-brmcoda}
\alias{log_posterior}
\alias{nuts_params}
\alias{rhat}
\alias{neff_ratio}
\alias{log_posterior.brmcoda}
\alias{nuts_params.brmcoda}
\alias{rhat.brmcoda}
\alias{neff_ratio.brmcoda}
\title{Extract Diagnostic Quantities from \code{brmsfit} Models in \code{brmcoda}}
\usage{
\method{log_posterior}{brmcoda}(object, ...)

\method{nuts_params}{brmcoda}(object, ...)

\method{rhat}{brmcoda}(x, ...)

\method{neff_ratio}{brmcoda}(object, ...)
}
\arguments{
\item{...}{Arguments passed to individual methods (if applicable).}

\item{x, object}{A \code{brmcoda} object or another \R object for which
the methods are defined.}
}
\value{
The exact form of the output depends on the method.
}
\description{
Extract Diagnostic Quantities from \code{brmsfit} Models in \code{brmcoda}
}
\seealso{
\code{\link[brms:log_posterior.brmsfit]{log_posterior.brmsfit}}

\code{\link[brms:nuts_params.brmsfit]{nuts_params.brmsfit}}

\code{\link[brms:rhat.brmsfit]{rhat.brmsfit}}

\code{\link[brms:neff_ratio.brmsfit]{neff_ratio.brmsfit}}
}
