% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{diagnostic-quantities}
\alias{diagnostic-quantities}
\alias{log_posterior}
\alias{nuts_params}
\alias{rhat}
\alias{neff_ratio}
\alias{log_posterior.brmcoda}
\alias{nuts_params.brmcoda}
\alias{rhat.brmcoda}
\alias{neff_ratio.brmcoda}
\title{Extract Diagnostic Quantities from \pkg{brmcoda} Models}
\usage{
\method{log_posterior}{brmcoda}(x)

\method{nuts_params}{brmcoda}(x)

\method{rhat}{brmcoda}(x)

\method{neff_ratio}{brmcoda}(x)
}
\arguments{
\item{x}{A \code{brmcoda} object.}

\item{...}{Arguments passed to individual methods.}
}
\value{
The exact form of the output depends on the method.
}
\description{
Extract Diagnostic Quantities from \pkg{brmcoda} Models
}
