% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{diagnostic_posterior.brmcoda}
\alias{diagnostic_posterior.brmcoda}
\title{Posteriors Sampling Diagnostic}
\usage{
\method{diagnostic_posterior}{brmcoda}(posteriors, diagnostic = c("ESS", "Rhat"), ...)
}
\arguments{
\item{posteriors}{An object of class \code{brmcoda}.}

\item{diagnostic}{Diagnostic metrics to compute.  Character (vector) or list
with one or more of these options: \code{"ESS"}, \code{"Rhat"}, \code{"MCSE"} or \code{"all"}.}

\item{...}{Other arguments passed to \code{\link{diagnostic_posterior}}.}
}
\description{
Extract diagnostic metrics (Effective Sample Size (\code{ESS}), \code{Rhat} and Monte
Carlo Standard Error \code{MCSE}).
}
\seealso{
\code{\link[bayestestR:diagnostic_posterior]{diagnostic_posterior}}
}
