% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Generic diagnostics function for \pkg{multilevelcoda}}
\usage{
diagnostics(x, ...)
}
\arguments{
\item{x}{An object to calculate diagnostics, primarily around extreme values
and assumed distributions.}

\item{...}{Additional arguments passed to methods.}
}
\value{
\code{diagnostics} class object
}
\description{
This is a generic function for diagnostics methods in \pkg{multilevelcoda}. See
\code{\link{diagnostics.complr}} for the method for \code{\link{complr}} objects.
}
\keyword{multivariate}
