% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.adj.subm}
\alias{.adj.subm}
\title{\subsection{Support Substitution Model}{Get substitution data
}}
\usage{
.adj.subm(sub, b, refg, subvar, iv, tmp, psi, substitute, i, j, k, ID = "1")
}
\description{
\subsection{Support Substitution Model}{Get substitution data
}
}
