% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get.bsubmargins.}
\alias{.get.bsubmargins.}
\title{Alternative model estimating AME for between-person substitutionwhich binds k level results for fitted()}
\usage{
.get.bsubmargins.(object, substitute, min, ysame, ...)
}
\description{
Alternative model estimating AME for between-person substitutionwhich binds k level results for fitted()
}
