% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get.bsubmargins}
\alias{.get.bsubmargins}
\title{AME for Between-person Substitution model}
\usage{
.get.bsubmargins(object, substitute, min, ysame, ...)
}
\description{
AME for Between-person Substitution model
}
