% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get.wsubmargins}
\alias{.get.wsubmargins}
\title{AME for Within-person substitution model}
\usage{
.get.wsubmargins(object, substitute, min, ysame, ...)
}
\description{
AME for Within-person substitution model
}
