% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{draws-index-brmcoda}
\alias{draws-index-brmcoda}
\alias{variables}
\alias{nvariables}
\alias{niterations}
\alias{nchains}
\alias{ndraws}
\alias{variables.brmcoda}
\alias{nvariables.brmcoda}
\alias{niterations.brmcoda}
\alias{nchains.brmcoda}
\alias{ndraws.brmcoda}
\title{#' Extract Posterior Draws
#'
#' Extract posterior draws in conventional formats
#' as data.frames, matrices, or arrays.
#'
#' @inheritParams brms::as.data.frame.brmsfit
#'
#' @return A data.frame, matrix, or array containing the posterior draws.
#'
#' @export
as.data.frame.brmcoda <- function(x, row.names = NULL, optional = TRUE, ...) {
as.data.frame(x$Model, ...)
}}
\usage{
\method{variables}{brmcoda}(x, ...)

\method{nvariables}{brmcoda}(x, ...)

\method{niterations}{brmcoda}(x)

\method{nchains}{brmcoda}(x)

\method{ndraws}{brmcoda}(x)
}
\arguments{
\item{x}{An object of class \code{brmcoda}.}

\item{...}{Arguments passed to individual methods.}
}
\description{
#' @rdname as.data.frame.brmcoda
#' @export
as.matrix.brmcoda <- function(x, ...) {
as.matrix(x$Model, ...)
}
}
\details{
#' @rdname as.data.frame.brmcoda
#' @export
as.array.brmcoda <- function(x, ...) {
as.array(x$Model, ...)
}
Index \code{brmcoda} objects
}
\seealso{
\code{\link[brms:variables]{variables.brmsfit}}

\code{\link[brms:nvariables.brmsfit]{nvariables.brmsfit}}

\code{\link[brms:niterations.brmsfit]{niterations.brmsfit}}

\code{\link[brms:nchains.brmsfit]{nchains.brmsfit}}

\code{\link[brms:ndraws.brmsfit]{ndraws.brmsfit}}
}
