% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmcoda.R
\name{emmcoda}
\alias{emmcoda}
\title{Estimated marginal means of composition.}
\usage{
emmcoda(object, specs, at, ...)
}
\arguments{
\item{specs}{A character vector specifying the names of the predictors over which EMMs are desired.}

\item{at}{A numeric vector passed to \code{\link{ref_grid}} specifying the predictor's values for EMMs.}

\item{...}{Further arguments passed to \code{\link{emmeans}}.}

\item{data}{A fitted \code{\link{mvmcoda}} model data. Required.}
}
\value{

}
\description{
This function obtains estimated marginal means (EMMs) of models
containing composition as the outcome
}
\examples{

emtest <- emmcoda(object = mvm1, specs = "STRESS", at = c(1:10))

}
