% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{examplefun}
\alias{examplefun}
\title{Example Function Title}
\usage{
examplefun(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The vector coerced to numeric vector in a data table.
}
\description{
This provides a few sentence description about the example function.
}
\examples{

examplefun(c(1L, 2L))

examplefun(TRUE)

examplefun(c(5.1, 2.2))
}
