% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitution-helpers.R
\name{get-substitution}
\alias{get-substitution}
\title{Helper functions used only internally to estimate substitution model}
\description{
Helper functions used only internally to estimate substitution model
}
