% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get.bsub}
\alias{get.bsub}
\title{Between-person Substitution model}
\usage{
get.bsub(
  object,
  substitute,
  mcomp,
  ysame,
  min,
  summary = summary,
  psi,
  ID = 1,
  cv = NULL,
  refg = NULL,
  ...
)
}
\description{
Between-person Substitution model
}
