% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complr-methods.R
\name{get.names.complr}
\alias{get.names.complr}
\title{Extract names from a \code{complr} object.}
\usage{
get.names.complr(object)
}
\arguments{
\item{complr}{A \code{complr} object}
}
\description{
Extract names from a \code{complr} object.
}
