% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_sbp}
\alias{get_sbp}
\title{Extract Sequential Binary Partition from a \code{complr} object.}
\usage{
get_sbp(object)
}
\arguments{
\item{object}{A \code{complr} object}
}
\description{
Extract Sequential Binary Partition from a \code{complr} object.
}
