% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_variables.complr}
\alias{get_variables.complr}
\alias{get_variables.brmcoda}
\title{Extract variable names from a \code{complr} object.}
\usage{
\method{get_variables}{complr}(object)

\method{get_variables}{brmcoda}(object)
}
\arguments{
\item{object}{A \code{brmcoda} object}
}
\description{
Extract variable names from a \code{complr} object.

Extract variable names from a \code{brmcoda} object.
}
