% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilr-methods.R
\name{is.compilr}
\alias{is.compilr}
\title{Checks if argument is a \code{compilr} object}
\usage{
is.compilr(x)
}
\arguments{
\item{x}{An object of class \code{compilr}.}
}
\description{
Checks if argument is a \code{compilr} object
}
