% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complr-methods.R
\name{is.complr}
\alias{is.complr}
\title{Checks if argument is a \code{complr} object}
\usage{
is.complr(x)
}
\arguments{
\item{x}{An object of class \code{complr}.}
}
\description{
Checks if argument is a \code{complr} object
}
