% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{is.diagnostics}
\alias{is.diagnostics}
\title{Checks if argument is a \code{diagnostics} object}
\usage{
is.diagnostics(x)
}
\arguments{
\item{x}{An object of class \code{diagnostics}.}
}
\description{
Checks if argument is a \code{diagnostics} object
}
