% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitution-helpers.R
\name{is.substitution}
\alias{is.substitution}
\title{Checks if argument is a \code{substitution} object}
\usage{
is.substitution(x)
}
\arguments{
\item{x}{An object of class \code{substitution}.}
}
\description{
Checks if argument is a \code{substitution} object
}
