% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{launch_shinystan.brmcoda}
\alias{launch_shinystan.brmcoda}
\alias{launch_shinystan}
\title{Interface to \pkg{shinystan}}
\usage{
\method{launch_shinystan}{brmcoda}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{brmcoda}.}

\item{...}{Optional arguments to pass to
\code{\link[shinystan:launch_shinystan]{launch_shinystan}} or
\code{\link[shiny:runApp]{runApp}}.}
}
\value{
An S4 shinystan object
}
\description{
Provide an interface to \pkg{shinystan} for models fitted with \pkg{brms}
}
\seealso{
\code{\link[shinystan:launch_shinystan]{launch_shinystan}}
}
