% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-methods.R
\name{loo.brmcoda}
\alias{loo.brmcoda}
\alias{loo}
\title{Efficient approximate leave-one-out cross-validation (LOO)}
\usage{
\method{loo}{brmcoda}(x, ...)
}
\arguments{
\item{x}{A \code{brmcoda} object.}

\item{...}{More \code{brmsfit} objects or further arguments
passed to the underlying post-processing functions.
In particular, see \code{\link[brms]{prepare_predictions}} for further
supported arguments.}
}
\value{
If just one object is provided, an object of class \code{loo}.
 If multiple objects are provided, an object of class \code{loolist}.
}
\description{
Perform approximate leave-one-out cross-validation based
on the posterior likelihood using the \pkg{loo} package.
For more details see \code{\link[loo:loo]{loo}}.
}
\seealso{
\code{\link[brms:loo.brmsfit]{loo.brmsfit}}
}
