% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m}
\alias{m}
\title{A fitted Bayesian multivariate multilevel compositional model using 'brmcoda'
An example of unadjusted 'brmcoda' object}
\format{
A list with 2 elemets.
\describe{
\item{CompILR}{A 'compilr' object}
\item{Model}{Fitted brm multivariate multilevel
mmodel with compositional outcome}
}
}
\usage{
m
}
\description{
A fitted Bayesian multivariate multilevel compositional model using 'brmcoda'
An example of unadjusted 'brmcoda' object
}
\keyword{datasets}
