% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcompd}
\alias{mcompd}
\title{Multilevel Compositional Data}
\format{
A data table containing 10 variables.
\describe{
\item{ID}{A unique identifier for each individual}
\item{Time}{Recurrence time of repeated measures by individual}
\item{Stress}{Self report stress measures on a 0 to 10 scale --- repeated measure}
\item{TST}{Total Sleep Time (minutes) --- repeated measure}
\item{WAKE}{Wake time while in bed, trying to sleep (minutes) --- repeated measure}
\item{MVPA}{Moderate to Vigorous Physical Activity (minutes) --- repeated measure}
\item{LPA}{Light Physical Activity (minutes) --- repeated measure}
\item{SB}{Sedentary Behavior (minutes) --- repeated measure}
\item{Age}{Age in years --- baseline measure only}
\item{Female}{Binary: whether participants identified as female (1) or not (0) --- baseline measure only}
}
}
\usage{
mcompd
}
\description{
A simulated dataset containing multiple days of compositional data.
}
\keyword{datasets}
