% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{multilevelcoda_sim}
\alias{multilevelcoda_sim}
\title{multilevelcoda Simulation Study Results}
\usage{
multilevelcoda_sim()
}
\value{
An S4 shiny object
}
\description{
Provide the full results for a simulation study testing the performance of \pkg{multilevelcoda}
}
