% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvcoda.R
\name{mvcoda}
\alias{mvcoda}
\title{A series of functions for single level multivariate compositional models}
\usage{
mvcoda(data, composition, sbp, formula, ...)
}
\arguments{
\item{formula}{A object of class \code{\link{brmsformula}}, \code{\link{mvbrmsformula}}:
A symbolic description of the model to be fitted.
Details of the model specification can be found in \code{\link{mvbrmsformula}}.}

\item{...}{Further arguments passed to \code{\link{brm}}.}

\item{compilr}{A \code{compilr} object containing data of composition, ILR coordinates,
and other variables used in the model.}
}
\value{

}
\description{
Fit Bayesian generalised (non-)linear multivariate compositional model
via full Bayesian inference using Stan,
when composition is the outcome.
}
\examples{
data(mcompd)
cilr <- compilr(data = mcompd, sbp = sbp, idvar = "ID")

## inspect names of ILR coordinates bfore passing to 'brm' model
names(cilr$BetweenILR)

## run mvcoda
mv1 <- mvcoda(compilr = cilr, formula = mvbind(ilr1, ilr2, ilr3, ilr4) ~ STRESS + Age + Female)

}
