% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvm}
\alias{mvm}
\title{A fitted Bayesian multivariate multilevel compositional model using 'mvmcoda'
An example of unadjusted 'mvmcoda' object}
\format{
A list with 2 elemets.
\describe{
\item{CompILR}{A 'compilr' object}
\item{Model}{Fitted brm multivariate multilevel
mmodel with compositional outcome}
}
}
\usage{
mvm
}
\description{
A fitted Bayesian multivariate multilevel compositional model using 'mvmcoda'
An example of unadjusted 'mvmcoda' object
}
\keyword{datasets}
