% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvmcoda.R
\name{mvmcoda}
\alias{mvmcoda}
\title{Fit Bayesian generalised (non-)linear multivariate multilevel compositional model
via full Bayesian inference using Stan,
when composition is the outcome.}
\usage{
mvmcoda(compilr, formula, ...)
}
\arguments{
\item{compilr}{A \code{compilr} object containing data of composition, ILR coordinates,
and other variables used in the model.}

\item{formula}{A object of class \code{\link{brmsformula}}, \code{\link{mvbrmsformula}}:
A symbolic description of the model to be fitted.
Details of the model specification can be found in \code{\link{mvbrmsformula}}.}

\item{...}{Further arguments passed to \code{\link{brm}}.}
}
\value{

}
\description{
This function utilises
}
\examples{
data(mcompd)
cilr <- compilr(data = mcompd, sbp = sbp, idvar = "ID")

## inspect names of ILR coordinates bfore passing to 'brm' model
names(cilr$BetweenILR)

## run mvmcoda
mvm1 <- mvmcoda(compilr = cilr, formula = mvbind(ilr1, ilr2, ilr3, ilr4) ~ STRESS + (1 | ID))

}
