% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complr-methods.R
\name{names.complr}
\alias{names.complr}
\title{Extract names from a \code{complr} object.}
\usage{
\method{names}{complr}(object)
}
\arguments{
\item{complr}{A \code{complr} object}
}
\description{
Extract names from a \code{complr} object.
}
