% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{nobs.brmcoda}
\alias{nobs.brmcoda}
\title{Extract Number of Observations from \pkg{brmcoda} Models}
\usage{
\method{nobs}{brmcoda}(x)
}
\arguments{
\item{object}{A \code{brmcoda} object.}
}
\description{
Extract Number of Observations from \pkg{brmcoda} Models
}
