% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diagnostics.R
\name{plot.diagnostics}
\alias{plot.diagnostics}
\title{Diagnostics Plot for Compositional Diagnostics}
\usage{
\method{plot}{diagnostics}(x, transform = c("clr", "raw"), ...)
}
\arguments{
\item{x}{A \code{diagnostics} class object created by
\code{\link{diagnostics.complr}}.}

\item{transform}{Character string indicating whether to plot the raw
composition values in \code{x$x} or centered logratio transformed values.
Defaults to \code{"clr"}.}

\item{...}{Currently unused.}
}
\value{
A ggplot graph object showing pairwise scatterplots for each
composition part, with point colour indicating extreme value status. Diagonal
panels show marginal densities and rug marks for observations flagged as
extreme values.
}
\description{
Make a scatterplot matrix of \code{\link{diagnostics}} results from a
\code{\link{complr}} object, with density and rug plots on the diagonal.
}
\examples{
data(mcompd)
data(sbp)

ids <- unique(mcompd$ID)[1:20]
cilr <- complr(
  data = mcompd[ID \%in\% ids, .SD[1:3], by = ID],
  sbp = sbp,
  parts = c("TST", "WAKE", "MVPA", "LPA", "SB"),
  idvar = "ID",
  total = 1440
)

between_diag <- diagnostics(cilr, level = "between", ev.perc = .05)

# Plot centered logratio transformed composition values
plot(between_diag)

# Plot on the raw composition scale
plot(between_diag, transform = "raw")
}
