% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoda.R
\name{plot.substitution}
\alias{plot.substitution}
\title{Substitution plot}
\usage{
\method{plot}{substitution}(object, to, ref, level, ...)
}
\arguments{
\item{object}{A \code{\link{substitution}} object containing the output of substitution models.}

\item{to}{A character value or vector specifying the names of the compositional parts
that were reallocated to in the model.}

\item{ref}{Either a character value or vector ((\code{grandmean} and/or \code{clustermean} or \code{users}),
Default to all \code{ref} available in the \code{\link{substitution}} object .}

\item{level}{A character string or vector (\code{between} and/or \code{within}).
Default to all \code{level} available in the \code{\link{substitution}} object.}

\item{...}{Further arguments passed to \code{\link{ggplot}}.}
}
\value{
A ggplot graph object showing the estimated difference in outcome when
each pair of compositional variables are substituted for a specific time.
}
\description{
Make a plot of \code{\link{substitution}} model results.
}
