% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoda.R
\name{plotcoda}
\alias{plotcoda}
\title{Generate marginal effects of composition plot}
\usage{
plotcoda(data, iv)
}
\arguments{
\item{data}{A dataset to use for plot.
It must be a \code{\link{mvmcoda}} or \code{\link{mvcoda}} object}

\item{iv}{A character string indicating name of the predictor variable.}

\item{compilr}{A \code{compilr} object containing data of composition, ILR coordinates,
and other variables used for plot.}

\item{dv}{A character string indicating the name of the compositional outcome variable.}

\item{font}{A character string indicating the name of user's preferred font. Default is \code{Times New Roman}.}

\item{...}{Further arguments passed to \code{ggplot}.}
}
\value{
A ggplot graph object showing the estimated difference in outcome when
each pair of compositional variables are substituted for a specific time.
}
\description{
This function is useful for visualising the
estimated changes in compositional outcome variable
associated with a specific change in some independent variable.
}
\examples{

data(mcompd)
plotsub(data = bsubctest2$TST, iv = "sleep")
}
