% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoda.R
\name{plotemmc}
\alias{plotemmc}
\title{Generate marginal effects of composition plot}
\usage{
plotemmc(data, x, y, font = "Times New Roman", ...)
}
\arguments{
\item{data}{A \code{\link{emmcoda}} object to use for plot.}

\item{x}{A character string specifying name of the predictor variable.}

\item{y}{A character string specifying the name of the compositional outcome variable.}

\item{font}{A character string specifying the name of user's preferred font. Default is \code{Times New Roman}.}

\item{...}{Further arguments passed to \code{\link{ggplot}}.}
}
\value{
A ggplot graph object showing the estimated difference in outcome when
each pair of compositional variables are substituted for a specific time.
}
\description{
This function is useful for visualising the
estimated changes in compositional outcome variable
associated with a specific change in some independent variable.
}
\examples{

data(mcompd)
plotemmc(data = emtest, x = "stress", y = "behaviours")
}
