% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoda.R
\name{plotsub}
\alias{plotsub}
\title{Generate substitution plot}
\usage{
plotsub(data, iv, dv, font = "Times New Roman", ...)
}
\arguments{
\item{data}{A dataset to use for plot.
It must be a component of a list resulted from one of the following functions:
\code{wsub}, \code{bsub}, \code{wsubmargins}, \code{bsubmargins}.}

\item{iv}{A character string indicating name of the compostional predictor variable.}

\item{dv}{A character string indicating the name of the outcome variable.}

\item{font}{A character string indicating the name of user's preferred font. Default is \code{Times New Roman}.}

\item{...}{Further arguments passed to \code{ggplot}.}
}
\value{
A ggplot graph object showing the estimated difference in outcome when
each pair of compositional variables are substituted for a specific time.
}
\description{
This function is useful for visualising the
estimated differences in outcomes when compositional variables
are substituted for a specific period of time.
}
\examples{

data(mcompd)
plotsub(data = bsubctest2$TST, iv = "sleep", dv = "stress")

}
