% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possub.R
\name{possub}
\alias{possub}
\title{Possible Pairwise Substitution}
\usage{
possub(count, data, idvar)
}
\arguments{
\item{count}{A numeric value indicating the number of variables within the composition.}

\item{data}{A dataset of composition and IDs. Ideally ID is the last column. Required.}

\item{idvar}{A string character indicating the name of the ID variable. Default is ID.}
}
\value{
A data table of all possible pairwise substitution.
}
\description{
This function generate a dataset of all possible pairwise substitution of a composition.
}
\examples{

data(mcompd)

posubtest <- possub(count = 5, data = mcompd[, 1:6], idvar = "ID")

str(posubtest)

## cleanup
rm(mcompd, posubtest)
}
