% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.brmcoda}
\alias{print.brmcoda}
\title{Print a summary for a fitted \code{brmsfit} model in a \code{brmcoda} object}
\usage{
\method{print}{brmcoda}(x, ...)
}
\arguments{
\item{x}{A \code{brmcoda} object.}

\item{...}{Additional arguments to be passed to to method \code{summary} of \code{brmsfit}.}
}
\description{
Print a summary for a fitted \code{brmsfit} model in a \code{brmcoda} object
}
\seealso{
\code{\link{summary.brmcoda}}
}
