% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.compilr}
\alias{print.compilr}
\title{Print a Summary for a \code{compilr} object}
\usage{
\method{print}{compilr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{compilr}.}

\item{...}{Other arguments passed to \code{summary.compilr}.}
}
\description{
Print a Summary for a \code{compilr} object
}
\seealso{
\code{\link{summary.compilr}}
}
