% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.complr}
\alias{print.complr}
\title{Print a Summary for a \code{complr} object}
\usage{
\method{print}{complr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{complr}.}

\item{...}{Other arguments passed to \code{\link{summary.complr}}.}
}
\description{
Print a Summary for a \code{complr} object
}
\examples{

cilr <- complr(data = mcompd, sbp = sbp,
                parts = c("TST", "WAKE", "MVPA", "LPA", "SB"),
                idvar = "ID")
print(cilr)
}
\seealso{
\code{\link{summary.complr}}
}
