% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-fit-prep.R
\name{print.mlsim_fit_prep}
\alias{print.mlsim_fit_prep}
\title{Print an Outcome Fit Preparation Object}
\usage{
\method{print}{mlsim_fit_prep}(x, ...)
}
\arguments{
\item{x}{An \code{mlsim_fit_prep} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an Outcome Fit Preparation Object
}
