% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.substitution}
\alias{print.substitution}
\title{Print a summary for a \code{substitution} object}
\usage{
\method{print}{substitution}(x, ...)
}
\arguments{
\item{x}{A \code{substitution} object.}

\item{...}{Additional arguments to be passed to to method \code{summary} of \code{substitution}.}
}
\description{
Print a summary for a \code{substitution} object
}
\seealso{
\code{\link{summary.substitution}}
}
