% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-print.R
\name{print.summary.mlsim_data}
\alias{print.summary.mlsim_data}
\title{Print a Simulated Data Summary}
\usage{
\method{print}{summary.mlsim_data}(x, ...)
}
\arguments{
\item{x}{A \code{summary.mlsim_data} object.}

\item{...}{Additional arguments passed to table printing.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Simulated Data Summary
}
