% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{prior_summary.brmcoda}
\alias{prior_summary.brmcoda}
\title{Extract Priors of a \code{brmsfit} from a \code{brmcoda} object}
\usage{
\method{prior_summary}{brmcoda}(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmcoda}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Compute Bayes factors from marginal likelihoods
}
\seealso{
\code{\link[brms:prior_summary.brmsfit]{prior_summary.brmsfit}}
}
