% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psub}
\alias{psub}
\title{Possible Pairwise Substitutions}
\format{
A data table containing 5 variables.
\describe{
\item{TST}{first compositional variable}
\item{WAKE}{second compositional variable}
\item{MVPA}{third compositional variable}
\item{LPA}{fourth compositional variable}
\item{SB}{fifth compositional variable}
}
}
\usage{
psub
}
\description{
A dataset containing possible pairwise subsitutions.
}
\keyword{datasets}
