% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{multilevelcoda Simulation Study results}
\format{
A list with 5 columns and 4 rows.
\describe{
\item{brmcoda_tab}{Simulation results for brmcoda() for tables}
\item{sub_tab}{Simulation results for substitution() for tables}
\item{brmcoda_plot}{Simulation results for brmcoda() for graphs}
\item{sub_plot}{Simulation results for substitution() for graphs}
}
}
\usage{
sim
}
\description{
A list of 4 components
}
\keyword{datasets}
