% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcoda.R
\name{subcoda}
\alias{subcoda}
\title{A series of functions for single level compositional models}
\usage{
subcoda(data, sbp, parts, substitute, formula, minute = 60, total = 1440, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table}
containing data of all variables used in the analysis.}

\item{sbp}{A signary matrix indicating sequential binary partition.}

\item{parts}{A character vector specifying the names of compositional variables.}

\item{substitute}{A \code{data.frame} or \code{data.table} of the possible substitution of variables.
This dataset can be computed using function \code{possub}.}

\item{formula}{A object of class \code{formula}, \code{brmsformula}:
A symbolic description of the model to be fitted.}

\item{minute}{A integer or numeric value indicating the maximum minute for which substitution model is desired.
Default to \code{60L} (i.e., the model loops through 1:60L minutes).}

\item{total}{A numeric value of the total amount to which the compositions should be closed. Default to \code{1440}.}

\item{...}{Further arguments passed to \code{\link{brm}}.}
}
\value{
A list with seven elements.
\itemize{
\item{\code{BrmsModel}}{ An object of class \code{brmsfit}, which contains the posterior draws
along with many other useful information about the model.}
\item{\code{SubstitutionResults}}{ A list containing the result of isotemporal substitution model.}
\item{\code{ILR}}{ Isometric log ratio transform of composition.}
\item{\code{Comp}}{A vector of class \code{acomp} representing one composition
or a matrix of class \code{acomp} representing multiple closed compositions each in one row.}
\item{\code{data}}{ The user's dataset or imputed dataset if the input data contains zeros.}
\item{\code{sbp}}{ The user-defined sequential binary partition matrix.}
\item{\code{parts}}{ Names of compositional variables.}
}
}
\description{
This function first computes composition and ILR coordinates,
then fits a Bayesian single-level compositional model to compositional predictors
and users' choices of outcomes and covariates,
and finally estimates the average marginal effects of
isotemporal compositional substitution model.
}
\examples{

data(mcompd)
data(sbp)

davg <- mcompd[, .(SB = mean(SB, na.rm = TRUE),
                   LPA = mean(LPA, na.rm = TRUE),
                   MVPA = mean(MVPA, na.rm = TRUE),
                   TST = mean(TST, na.rm = TRUE),
                   WAKE = mean(WAKE, na.rm = TRUE),
                   STRESS = mean(STRESS, na.rm = TRUE),
                   Age = mean (Age, na.rm = TRUE),
                   Female = na.omit(Female)[1]),
                   by = .(ID)]
ps <- possub(parts = c("TST", "WAKE", "MVPA", "LPA", "SB")

scoda <- subcoda(data = davg, parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), 
                       sbp = sbp, minute = 10, substitute = ps,
                       formula = STRESS ~ ilr1 + ilr2 + ilr3 + ilr4)
 
 rm(mcompd, sbp, davg, scoda)
}
