% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot.R
\name{subplot}
\alias{subplot}
\title{Generate multilevel substitution plot}
\usage{
subplot(object, predictor, outcome)
}
\arguments{
\item{object}{A dataset resulted from the substitution model.}

\item{predictor}{A name of the compostional predictor variable
that is substituted to/from. # nolint}

\item{outcome}{A name of the outcome variable.}
}
\value{
A ggplot graph object
}
\description{
This function is designed to generate a plot showing the
estimated differences in outcomes when compositional variables
are substituted for a specific period of time
}
\examples{
## TODO
}
