% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitution.R
\name{substitution}
\alias{substitution}
\title{Between-person Substitution Model.}
\usage{
substitution(
  object,
  delta,
  regrid = NULL,
  summary = TRUE,
  level = c("between", "within"),
  type = c("conditional", "marginal"),
  ...
)
}
\arguments{
\item{object}{A fitted \code{\link{brmcoda}} object.}

\item{delta}{A integer, numeric value or vector indicating the amount of change in compositional parts
for substitution.}

\item{regrid}{If non-\code{NULL}, a \code{data.table} of reference grid consisting
of combinations of covariates over which predictions are made.
Otherwise, the reference grid is constructed via \code{\link{ref_grid}}.}

\item{summary}{A logical value.
Should the estimate at each level of the reference grid (\code{FALSE})
or their average (\code{TRUE}) be returned? Default to \code{TRUE}.}

\item{...}{Additional arguments to be passed to \code{\link{describe_posterior}}.}
}
\value{
A list containing the result of isotemporal multilevel substitution model.
Each elements of the list is the substitution estimation for a compositional variables,
which include at least six elements.
\itemize{
\item{\code{Mean}}{ Posterior means.}
\item{\code{CI_low}} and \item{\code{CI_high}}{ 95\% credible intervals.}
\item{\code{Delta}}{ Amount substituted across compositional parts.}
\item{\code{From}}{ Compositional variable that is substituted from.}
\item{\code{To}}{ Compositional variables that is substituted to.}
\item{\code{Level}}{Level where changes in composition takes place.}
\item{\code{EffectType}}{Either estimated \code{conditional} or average \code{marginal} changes.}
}
}
\description{
Estimate the difference in outcomes
when compositional variables are substituted for a specific unit of change.
The resulting \code{substitution} encapsulates
substitution estimation across all compositional variables
present in the \code{\link{brmcoda}} object.
}
\examples{
\dontrun{
data(mcompd)
data(sbp)
data(psub)
cilr <- compilr(data = mcompd, sbp = sbp, 
                parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), idvar = "ID")

# model with compositional predictor at between and between-person levels
m <- brmcoda(compilr = cilr, 
             formula = STRESS ~ bilr1 + bilr2 + bilr3 + bilr4 + 
                                wilr1 + wilr2 + wilr3 + wilr4 + (1 | ID), 
             chain = 1, iter = 500)
             
subm <- substitution(object = m, delta = c(30, 60),
                     type = "marginal", level = "within")
}
}
