% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.brmcoda}
\alias{summary.brmcoda}
\title{Create a summary of a fitted \code{brmsfit} model in a \code{brmcoda} object}
\usage{
\method{summary}{brmcoda}(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmcoda}.}

\item{...}{Other arguments passed to \code{summary.brmsfit}}
}
\description{
Create a summary of a fitted \code{brmsfit} model in a \code{brmcoda} object
}
