% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.substitution}
\alias{summary.substitution}
\title{Summary \code{\link{substitution}}}
\usage{
\method{summary}{substitution}(object, delta, to, from, ref, level, digits = 2, ...)
}
\arguments{
\item{object}{A \code{\link{substitution}} object containing the output of substitution models.}

\item{delta}{A integer, numeric value or vector indicating the desired \code{delta}
at which substitution results should be summarised.
Default to all \code{delta} available in the \code{\link{substitution}} object.}

\item{to}{A character value or vector specifying the names of the compositional parts
that were reallocated to in the model.}

\item{from}{A character value or vector specifying the names of the compositional parts
that were reallocated from in the model.}

\item{ref}{Either a character value or vector ((\code{grandmean} and/or \code{clustermean} or \code{users}),
Default to all \code{ref} available in the \code{\link{substitution}} object .}

\item{level}{A character string or vector (\code{between} and/or \code{within}).
Default to all \code{level} available in the \code{\link{substitution}} object .}

\item{digits}{A integer value used for number formatting. Default to 3.}

\item{...}{generic argument, not in use.}
}
\value{
A summary of \code{\link{substitution}} object with twelve elements.
\itemize{
\item{\code{BetweenComp}}{ A vector of class \code{acomp} representing one closed between-person composition
or a matrix of class \code{acomp} representing multiple closed between-person compositions each in one row.}
\item{\code{WithinComp}}{ A vector of class \code{acomp} representing one closed within-person composition
or a matrix of class \code{acomp} representing multiple closed within-person compositions each in one row.}
\item{\code{TotalComp}}{ A vector of class \code{acomp} representing one closed total composition
or a matrix of class \code{acomp} representing multiple closed total compositions each in one row.}
\item{\code{BetweenILR}}{ Isometric log ratio transform of between-person composition.}
\item{\code{WithinILR}}{ Isometric log ratio transform of within-person composition.}
\item{\code{TotalILR}}{ Isometric log ratio transform of total composition.}
\item{\code{data}}{ The user's dataset or imputed dataset if the input data contains zeros.}
\item{\code{psi}}{ A ILR matrix associated with user-defined partition structure.}
\item{\code{sbp}}{ The user-defined sequential binary partition matrix.}
\item{\code{parts}}{ Names of compositional variables.}
\item{\code{idvar}}{ Name of the variable containing IDs.}
\item{\code{total}}{ Total amount to which the compositions is closed.}
}
}
\description{
This method allows for summarising an existing \code{\link{substitution}} object.
}
