% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilr.R
\name{tinycompilr}
\alias{tinycompilr}
\title{Compute useful indices from a (dataset of) single-level composition(s)}
\usage{
tinycompilr(data, sbp, parts, total = 1440)
}
\description{
Compute useful indices from a (dataset of) single-level composition(s)
}
