% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.brmcoda}
\alias{update.brmcoda}
\title{Update \code{\link{brmcoda}} models}
\usage{
\method{update}{brmcoda}(object, newcilr = NULL, newdata = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{\link{brmcoda}} object to be updated. Required.}

\item{newcilr}{A \code{\link{compilr}} object containing data of composition,
ILR coordinates, and other variables used in the updated model.}

\item{newdata}{A \code{data.frame} or \code{data.table}
containing data of all variables used in the analysis.
It must include a composition and the same ID variable as the existing \code{\link{compilr}} object.}

\item{...}{Further arguments passed to \code{\link{update}}.}
}
\description{
This method allows for updating an existing \code{\link{brmcoda}} object.
}
\details{
Details of the model specification can be found in \code{\link{brmsformula}}.
}
\examples{
\dontrun{
data(mcompd)
data(sbp)
cilr <- compilr(data = mcompd, sbp = sbp, 
        parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), idvar = "ID")

# model with compositional predictor at between and within-person levels
fit <- brmcoda(compilr = cilr, 
              formula = STRESS ~ bilr1 + bilr2 + bilr3 + bilr4 +
                                 wilr1 + wilr2 + wilr3 + wilr4 + (1 | ID), 
              chain = 1, iter = 500,
              backend = "cmdstanr")

newdat <- mcompd[ID != 1] # excluding ID 1
fit_new <- update(fit, newdata = newdat)
}
}
