% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsubmargins.R
\name{wsubmargins}
\alias{wsubmargins}
\title{Between-person Marginal Substitution Model.}
\usage{
wsubmargins(object, base, delta, level, type, ...)
}
\arguments{
\item{object}{A \code{\link{brmcoda}} object. Required.}

\item{base}{A \code{data.frame} or \code{data.table} of the possible substitution of variables.
This dataset can be computed using function \code{possub}. Required.}

\item{delta}{A integer, numeric value or vector indicating the amount of change in compositional parts
for substitution.}

\item{...}{Additional arguments to be passed to \code{\link{describe_posterior}}.}
}
\value{
A list containing the result of isotemporal multilevel substitution model.
Each elements of the list is the substitution estimation for a compositional variables,
which include at least six elements.
\itemize{
\item{\code{Mean}}{ Posterior means.}
\item{\code{CI_low}} and \item{\code{CI_high}}{ 95\% credible intervals.}
\item{\code{MinSubstituted}}{ Minute substituted within the composition.}
\item{\code{Substitute}}{Compositional variables to be substituted from/to.}
\item{\code{Predictor}}{Central compositional variable to be substituted from/to.}
}
}
\description{
Using a fitted model object, estimates the the average marginal difference
when compositional variables are substituted for a specific amount
at \code{within-person} level. The resulting \code{wsubmargins} encapsulates
substitution estimation across all compositional variables present
in the \code{\link{brmcoda}} object.
}
\examples{
\dontrun{
data(mcompd)
data(sbp)
data(psub)
cilr <- compilr(data = mcompd, sbp = sbp, 
                parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), idvar = "ID")
                
library(doFuture)
registerDoFuture()
plan(multisession, workers = 5)

# model with compositional predictor at between and within-person levels
m <- brmcoda(compilr = cilr, 
             formula = STRESS ~ bilr1 + bilr2 + bilr3 + bilr4 + 
                                wilr1 + wilr2 + wilr3 + wilr4 + (1 | ID), 
             chain = 1, iter = 500)             
subm <- wsubmargins(object = m, base = psub, delta = 5)
}
}
