% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilr-methods.R
\name{as.data.frame.compilr}
\alias{as.data.frame.compilr}
\alias{as.matrix.compilr}
\title{Extract Compositional Data from \code{compilr} object.}
\usage{
\method{as.data.frame}{compilr}(
  x,
  row.names = NULL,
  optional = TRUE,
  class = c("composition", "logratio"),
  level = c("between", "within", "total"),
  ...
)

\method{as.matrix}{compilr}(
  x,
  class = c("composition", "logratio"),
  level = c("between", "within", "total"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{compilr}.}

\item{row.names, optional}{Unused and only added for consistency with
the \code{\link[base:as.data.frame]{as.data.frame}} generic.}

\item{class}{Optional. Can be \code{"composition"} and/or \code{"logratio"} to
specify the geometry of the composition.}

\item{level}{Optional. Can be \code{"between"}, \code{"within"}, and/or \code{total}
indicating the level of the geometry.}

\item{...}{generic argument, not in use.}
}
\description{
Extract amounts and compositions in conventional formats
as data.frames, matrices, or arrays.
}
