% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{bayes_factor.brmcoda}
\alias{bayes_factor.brmcoda}
\title{Bayes Factors from Marginal Likelihoods}
\usage{
\method{bayes_factor}{brmcoda}(x1, x2, ...)
}
\arguments{
\item{x1}{A \code{brmcoda} object.}

\item{x2}{Another \code{brmcoda} object based on the same responses.}

\item{...}{Other arguments passed to \code{\link[brms:bayes_factor.brmsfit]{bayes_factor.brmsfit}}.}
}
\description{
Compute Bayes factors from marginal likelihoods
}
\seealso{
\code{\link[brms:bayes_factor.brmsfit]{bayes_factor.brmsfit}}
}
