% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsub.R
\name{bsub}
\alias{bsub}
\title{Between-person Substitution Model (from sample's compositional mean)}
\usage{
bsub(data, substitute, minute = 60)
}
\arguments{
\item{data}{(2) relevant inputs for substitution model.}

\item{substitute}{A data frame or data table indicating the possible substitution of variables. This dataset can be computed using \code{possub}. Required.}

\item{minute}{A integer or numeric value indicating the minute that compositional variable are substituted to/from. Default is 60L.}
}
\value{
A list
}
\description{
Estimate the difference in outcomes
when compositional variables are substituted for a specific time period.
at between-person level.
}
\examples{
## TODO

sbp <- matrix(c(
-1, -1, -1,-1, 1,
1, -1, -1, -1, 0,
0, 1, -1, -1, 0,
0, 0, 1, -1, 0), ncol = 5, byrow = TRUE)

data(mcompd)

posubtest <- possub(count = 5, data = mcompd[, 1:6], idvar = "ID")

bsubtest <- bsub(data = brmcodatest, substitute = posubtest, minute = 10)

## cleanup
rm(bsubtest, mcompd)
}
