% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsub.R
\name{bsub}
\alias{bsub}
\title{Between-person Substitution Model (from sample's compositional mean)}
\usage{
bsub(data, substitute, minute = 60L)
}
\arguments{
\item{data}{A \code{brmcoda} object that contains results of a fitted \code{brm} model and
relevant inputs of substitution model from \code{compilr}.}

\item{substitute}{A \code{data.frame} or \code{data.table} of the possible substitution of variables.
This dataset can be computed using function \code{possub}. Required.}

\item{minute}{A integer or numeric value specifying the minute that
compositional variable are substituted to/from. Default is \code{60L}.}
}
\value{
A list of results from substitution models for all compositional variables.
}
\description{
Estimate the difference in outcomes
when compositional variables are substituted for a specific time period.
at between-person level.
}
\examples{
## TODO


data(mcompd)
data(sbp)
ps <- possub(data = mcompd, composition = c("TST", "WAKE", "MVPA", "LPA", "SB"))

bsubtest <- bsub(data = mcm, substitute = ps, minute = 10)

## cleanup
rm(bsubtest, mcompd)
}
