% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsubmargins.R
\name{bsubmargins}
\alias{bsubmargins}
\title{Estimating Average Marginal Effects for Between-person Isotemporal Substitution Model.}
\usage{
bsubmargins(object, substitute, minute = 60L, ...)
}
\arguments{
\item{object}{A \code{\link{brmcoda}} object. Required.}

\item{substitute}{A data frame or data table indicating the possible substitution of variables.
This dataset can be computed using \code{possub}. Required.}

\item{minute}{A integer or numeric value indicating the maximum minute for which substitution model is desired.
Default to 60L. In this case, the model loops through 1:60L minutes.}

\item{...}{Additional arguments to be passed to \code{\link{describe_posterior}}.}
}
\value{
A list containing the result of isotemporal multilevel substitution model.
Each elements of the list is the substitution estimation for a compositional variables,
which include six or seven elements.
\itemize{
\item{\code{Mean}}{ Posterior means.}
\item{\code{CI_low}} and \item{\code{CI_high}}{ 95\% credible intervals.}
\item{\code{MinSubstituted}}{ Minute substituted within the composition.}
\item{\code{Substitute}}{Compositional variables to be substituted from/to.}
\item{\code{Predictor}}{Central compositional variable to be substituted from/to.}
}
}
\description{
Using a fitted model object, estimates the difference in outcomes
when compositional variables are substituted for a specific period
at between-person level. The resulting \code{bsubmargins} encapsulates
substitution estimation across all compositional variables present
in the \code{\link{brmcoda}} object.
}
\examples{
ps <- possub(data = mcompd, composition = c("TST", "WAKE", "MVPA", "LPA", "SB"))

library(doFuture)
registerDoFuture()
plan(multisession, workers = 5)
system.time(bsmtest2 <- bsubmargins(object = adjbrmcodatest, substitute = ps, minute = 5))
}
