% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{bsubmargins}
\alias{bsubmargins}
\title{Margianl Effects of Between-person Isotemporal Substitution.}
\usage{
bsubmargins(data, substitute, minute = 60L)
}
\arguments{
\item{data}{A fitted \code{\link{brmcoda}} model data. Required.}

\item{substitute}{A data frame or data table indicating the possible substitution of variables.
This dataset can be computed using \code{possub}. Required.}

\item{minute}{A integer or numeric value indicating the maximum minute for which substitution model is desired.
Default to 60L. In this case, the model loops through 1:60L minutes.}
}
\value{
A list containing the result of isotemporal multilevel substitution model.
Each elements of the list corresspond to a compositional variable.
}
\description{
This function estimates the difference in outcomes
when compositional variables are substituted for a specific period
at between-person level. The model loops through all compositional variables present
in the \code{\link{brmcoda}} object.
}
\examples{
ps <- possub(data = mcompd, composition = c("TST", "WAKE", "MVPA", "LPA", "SB"))

bsmtestc <- bsubmargins(data = mcm, substitute = ps, minute = 2)
}
