% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsubmargins.R
\name{bsubmargins}
\alias{bsubmargins}
\title{Between-person Marginal Substitution Model.}
\usage{
bsubmargins(object, base, delta, level = "between", type = "marginal", ...)
}
\arguments{
\item{object}{A \code{\link{brmcoda}} object.}

\item{base}{A \code{data.frame} or \code{data.table} of possible substitution of variables.
This dataset can be computed using function \code{possub}. Required.}

\item{delta}{A integer, numeric value or vector indicating the amount of change in compositional parts
for substitution.}

\item{...}{Additional arguments to be passed to \code{\link{describe_posterior}}.}
}
\value{
A list containing the result of isotemporal multilevel substitution model.
Each elements of the list is the substitution estimation for a compositional variables,
which include at least six elements.
\itemize{
\item{\code{Mean}}{ Posterior means.}
\item{\code{CI_low}} and \item{\code{CI_high}}{ 95\% credible intervals.}
\item{\code{MinSubstituted}}{ Minute substituted within the composition.}
\item{\code{Substitute}}{Compositional variables to be substituted from/to.}
\item{\code{Predictor}}{Central compositional variable to be substituted from/to.}
}
}
\description{
Using a fitted model object, estimates the average marginal difference
when compositional variables are substituted for a specific amount
at \code{between-person} level. The resulting \code{bsubmargins} encapsulates
substitution estimation across all compositional variables present
in the \code{\link{brmcoda}} object.
}
\examples{
\dontrun{
data(psub)
data(mcompd)
data(psub)
cilr <- compilr(data = mcompd[ID \%in\% 1:10, .SD[1:3], by = ID], sbp = sbp, 
                parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), idvar = "ID")

m <- brmcoda(compilr = cilr, 
             formula = STRESS ~ bilr1 + bilr2 + bilr3 + bilr4 + wilr1 + 
             wilr2 + wilr3 + wilr4 + Female + (1 | ID), chains = 1, iter = 500)
               
subm <- bsubmargins(object = m, base = psub, delta = 5)
}
}
