% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsubstitution-margins.R
\name{bsubmargins}
\alias{bsubmargins}
\title{Margianl Effects of Between-person Substitution

Estimate the difference in outcomes
when compositional variables are substituted for a specific time period
at between-person level.}
\usage{
bsubmargins(data, substitute, minute = 60L)
}
\arguments{
\item{data}{A fitted \code{brms} model data. Required.}

\item{substitute}{A data frame or data table indicating the possible substitution of variables. This dataset can be computed using \code{possub}. Required.}

\item{minute}{A integer or numeric value indicating the minute that compositional variable are substituted to/from. Default is 60L.}
}
\value{
A list
}
\description{
Margianl Effects of Between-person Substitution

Estimate the difference in outcomes
when compositional variables are substituted for a specific time period
at between-person level.
}
\examples{
ps <- possub(data = mcompd, composition = c("TST", "WAKE", "MVPA", "LPA", "SB"))

bsubmarginstest <- bsubmargins(data = mcmc, substitute = ps, minute = 10)
}
