% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complr.R
\name{compilr}
\alias{compilr}
\alias{complr}
\title{Indices from a (dataset of) Multilevel Composition(s)}
\usage{
compilr(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{complr}}.}
}
\value{
A \code{\link{complr}} object with at least 12 elements.
\item{\code{comp}}{ A vector of class \code{acomp} representing one closed composition
or a matrix of class \code{acomp} representing multiple closed  compositions each in one row.}
\item{\code{between_comp}}{ A vector of class \code{acomp} representing one closed between-person composition
or a matrix of class \code{acomp} representing multiple closed between-person compositions each in one row.}
\item{\code{within_comp}}{ A vector of class \code{acomp} representing one closed within-person composition
or a matrix of class \code{acomp} representing multiple closed within-person compositions each in one row.}
\item{\code{logratio}}{ Log ratio transform of composition.}
\item{\code{between_logratio}}{ Log ratio transform of between-person composition.}
\item{\code{within_logratio}}{ Log ratio transform of within-person composition.}
\item{\code{data}}{ The user's dataset or imputed dataset if the input data contains zeros.}
\item{\code{transform}}{ Type of transform applied on compositional data.}
\item{\code{parts}}{ Names of compositional variables.}
\item{\code{idvar}}{ Name of the variable containing IDs.}
\item{\code{shape}}{ The shape of the input dataset.}
\item{\code{total}}{ Total amount to which the compositions is closed.}
}
\description{
Compute sets of compositions and log ratio transformation for multilevel compositional data
}
