% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complr.R
\name{compilr}
\alias{compilr}
\title{Indices from a (dataset of) Multilevel Composition(s) (deprecated.)}
\usage{
compilr(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{complr}}.}
}
\value{
A \code{\link{complr}} object with at least the following elements.
\item{\code{X}}{ A vector of class \code{acomp} representing one closed composition
or a matrix of class \code{acomp} representing multiple closed  compositions each in one row.}
\item{\code{bX}}{ A vector of class \code{acomp} representing one closed between-person composition
or a matrix of class \code{acomp} representing multiple closed between-person compositions each in one row.}
\item{\code{wX}}{ A vector of class \code{acomp} representing one closed within-person composition
or a matrix of class \code{acomp} representing multiple closed within-person compositions each in one row.}
\item{\code{Z}}{ Log ratio transform of composition.}
\item{\code{bZ}}{ Log ratio transform of between-person composition.}
\item{\code{wZ}}{ Log ratio transform of within-person composition.}
\item{\code{data}}{ The user's dataset or imputed dataset if the iiut data contains zeros.}
\item{\code{transform}}{ Type of transform applied on compositional data.}
\item{\code{parts}}{ Names of compositional variables.}
\item{\code{idvar}}{ Name of the variable containing IDs.}
\item{\code{total}}{ Total amount to which the compositions is closed.}
}
\description{
Indices from a (dataset of) Multilevel Composition(s) (deprecated.)
}
\seealso{
\code{\link{complr}}
}
