% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitution-helpers.R
\name{create_substitution}
\alias{create_substitution}
\title{Constructor function for \code{substitution} class.}
\usage{
create_substitution(
  BetweenSub,
  WithinSub,
  BetweenSubMargins,
  WithinSubMargins,
  delta,
  ref,
  level,
  weight,
  parts,
  summary
)
}
\arguments{
\item{BetweenSub}{A list of results from \code{bsub} or \code{NULL}}

\item{WithinSub}{A list of results from \code{wsub} or \code{NULL}}

\item{BetweenSubMargins}{A list of results from \code{bsubmargins} or \code{NULL}}

\item{WithinSubMargins}{A list of results from \code{wsubmargins} or \code{NULL}}

\item{delta}{A numeric vector of the amount of substitution}

\item{ref}{A character value specifying the reference grid}

\item{level}{A character value specifying the level of substitution}

\item{weight}{The weight to use in calculation of the reference composition}

\item{parts}{The parts of the composition}

\item{summary}{A logical value specifying whether to summarize the results}
}
\value{
An object of class \code{substitution}
}
\description{
Constructor function for \code{substitution} class.
}
\seealso{
\code{\link{substitution}}
}
