% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitution-helpers.R
\name{create_substitution}
\alias{create_substitution}
\title{Constructor function for \code{substitution} class.}
\usage{
create_substitution(
  between_simple_sub,
  within_simple_sub,
  simple_sub,
  between_avg_sub,
  within_avg_sub,
  avg_sub,
  delta,
  ref,
  level,
  weight,
  parts,
  aorg
)
}
\arguments{
\item{between_simple_sub}{A list of results from \code{bsub} or \code{NULL}}

\item{within_simple_sub}{A list of results from \code{wsub} or \code{NULL}}

\item{simple_sub}{A list of results from \code{sub} or \code{NULL}}

\item{between_avg_sub}{A list of results from \code{bsubmargins} or \code{NULL}}

\item{within_avg_sub}{A list of results from \code{wsubmargins} or \code{NULL}}

\item{avg_sub}{A list of results from \code{submargins} or \code{NULL}}

\item{delta}{A numeric vector of the amount of substitution}

\item{ref}{A character value specifying the reference grid}

\item{level}{A character value specifying the level of substitution}

\item{weight}{The weight to use in calculation of the reference composition}

\item{parts}{The parts of the composition}

\item{aorg}{A logical value specifying whether to summarize the results (average over the reference grid)}
}
\value{
An object of class \code{substitution}
}
\description{
Constructor function for \code{substitution} class.
}
\seealso{
\code{\link{substitution}}
}
