% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{draws-index-brmcoda}
\alias{draws-index-brmcoda}
\alias{variables.brmcoda}
\alias{nvariables.brmcoda}
\alias{niterations.brmcoda}
\alias{nchains.brmcoda}
\alias{ndraws.brmcoda}
\title{Index \code{brmcoda} objects}
\usage{
\method{variables}{brmcoda}(x, ...)

\method{nvariables}{brmcoda}(x, ...)

\method{niterations}{brmcoda}(x)

\method{nchains}{brmcoda}(x)

\method{ndraws}{brmcoda}(x)
}
\arguments{
\item{x}{An object of class \code{brmcoda}.}
}
\description{
Index \code{brmcoda} objects
}
