% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{draws-index-brmcoda}
\alias{draws-index-brmcoda}
\alias{variables}
\alias{nvariables}
\alias{niterations}
\alias{nchains}
\alias{ndraws}
\alias{variables.brmcoda}
\alias{nvariables.brmcoda}
\alias{niterations.brmcoda}
\alias{nchains.brmcoda}
\alias{ndraws.brmcoda}
\title{Index \code{brmcoda} objects}
\usage{
\method{variables}{brmcoda}(x, ...)

\method{nvariables}{brmcoda}(x, ...)

\method{niterations}{brmcoda}(x)

\method{nchains}{brmcoda}(x)

\method{ndraws}{brmcoda}(x)
}
\arguments{
\item{x}{An object of class \code{brmcoda}.}

\item{...}{Arguments passed to individual methods.}
}
\description{
Index \code{brmcoda} objects
}
\seealso{
\code{\link[brms:variables]{variables.brmsfit}}

\code{\link[brms:nvariables.brmsfit]{nvariables.brmsfit}}

\code{\link[brms:niterations.brmsfit]{niterations.brmsfit}}

\code{\link[brms:nchains.brmsfit]{nchains.brmsfit}}

\code{\link[brms:ndraws.brmsfit]{ndraws.brmsfit}}
}
