% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmcoda.R
\name{emmcoda}
\alias{emmcoda}
\title{Estimated marginal means of composition.}
\usage{
emmcoda(object, x = NULL, at = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{\link{mvmcoda}} model data. Required.}

\item{x}{A character vector specifying the names of the predictors over which EMMs are desired.
If \code{NULL}, the average of all predictionu is returned. Default to \code{NULL}.}

\item{at}{A numeric value, vector or list passed to \code{\link{ref_grid}}
specifying the predictor's values for EMMs.}

\item{...}{Additional arguments to be passed to \code{\link{describe_posterior}}.}
}
\value{
A list with four elements.
}
\description{
This function obtains estimated marginal means (EMMs) of models
containing composition as the outcome
A sample from the posterior distribution of the regression coefficients.
}
