% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{fitted.brmcoda}
\alias{fitted.brmcoda}
\title{Expected Values of the Posterior Predictive Distribution}
\usage{
\method{fitted}{brmcoda}(
  object,
  newdata = NULL,
  re_formula = NULL,
  scale = c("response", "linear", "comp"),
  summary = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{brmcoda}.}

\item{newdata}{An optional data.frame for which to evaluate predictions. If
\code{NULL} (default), the original data of the model is used.
\code{NA} values within factors are interpreted as if all dummy
variables of this factor are zero. This allows, for instance, to make
predictions of the grand mean when using sum coding.}

\item{re_formula}{formula containing group-level effects to be considered in
the prediction. If \code{NULL} (default), include all group-level effects;
if \code{NA}, include no group-level effects.}

\item{scale}{Either \code{"response"} or \code{"linear"} or \code{"comp"}.
If either \code{"response"} or \code{"linear"},
results are returned using the corresponding method in \code{predict.brmsfit}.
\code{"comp"} is only relevant to multivariate models of class \code{mvbrmsformula}
(i.e., when response is composition). If \code{"comp"} is specified,
results are returned on the compositional scale of the response variable.}

\item{summary}{Should summary statistics be returned instead of the raw values?
Default is \code{TRUE}.}

\item{...}{Further arguments passed to \code{\link{fitted.brmsfit}}
that control additional aspects of prediction.}
}
\description{
Compute posterior draws of the expected value of the posterior predictive
distribution. Can be performed for the data used to fit the model (posterior
predictive checks) or for new data. By definition, these predictions have
smaller variance than the posterior predictions performed by the
\code{\link{posterior_predict.brmsfit}} method. This is because only the
uncertainty in the expected value of the posterior predictive distribution is
incorporated in the draws computed by \code{posterior_epred} while the
residual error is ignored there. However, the estimated means of both methods
averaged across draws should be very similar.
}
