% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R, R/complr-methods.R
\name{get_variables.brmcoda}
\alias{get_variables.brmcoda}
\alias{get_variables}
\alias{get_variables.complr}
\title{Extract variable names from a \code{brmcoda} object.}
\usage{
\method{get_variables}{brmcoda}(object)

\method{get_variables}{complr}(object)

\method{get_variables}{complr}(object)
}
\arguments{
\item{object}{A \code{complr} object}
}
\description{
Extract variable names from a \code{brmcoda} object.

Extract variable names from a \code{complr} object.
}
