% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_variables}
\alias{get_variables}
\alias{get_variables.complr}
\alias{get_variables.brmcoda}
\title{Extract variable names from an object}
\usage{
get_variables(object)

\method{get_variables}{complr}(object)

\method{get_variables}{brmcoda}(object)
}
\arguments{
\item{object}{A \code{brmcoda} object}
}
\value{
A list of variable names.
}
\description{
Generic function to extract variable names from a supported object.
}
\examples{
# For a complr object:
# get_variables(complr_object)

# For a brmcoda object:
# get_variables(brmcoda_object)

}
