% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{is.brmcoda}
\alias{is.brmcoda}
\title{Checks if argument is a \code{brmcoda} object}
\usage{
is.brmcoda(x)
}
\arguments{
\item{x}{An object of class \code{brmcoda}.}
}
\description{
Checks if argument is a \code{brmcoda} object
}
