% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{loo.brmcoda}
\alias{loo.brmcoda}
\alias{loo}
\title{Efficient approximate leave-one-out cross-validation (LOO)}
\usage{
\method{loo}{brmcoda}(x, ...)
}
\arguments{
\item{x}{A \code{brmcoda} object.}

\item{...}{More \code{brmsfit} objects or further arguments
passed to the underlying post-processing functions.
In particular, see \code{\link[brms]{prepare_predictions}} for further
supported arguments.}
}
\value{
If just one object is provided, an object of class \code{loo}.
 If multiple objects are provided, an object of class \code{loolist}.
}
\description{
Perform approximate leave-one-out cross-validation based
on the posterior likelihood using the \pkg{loo} package.
For more details see \code{\link[loo:loo]{loo}}.
}
\examples{
\donttest{
if(requireNamespace("cmdstanr")){
  m <- brmcoda(compilr = compilr(data = mcompd, sbp = sbp,
                                 parts = c("TST", "WAKE", "MVPA", "LPA", "SB"),
                                 idvar = "ID", total = 1440),
  formula = Stress ~ bilr1 + bilr2 + bilr3 + bilr4 +
    wilr1 + wilr2 + wilr3 + wilr4 + (1 | ID),
  chain = 1, iter = 500,
  backend = "cmdstanr")
  
  (loo1 <- loo(m))
}}
}
\seealso{
\code{\link[brms:loo.brmsfit]{loo.brmsfit}}
}
