% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{model.frame.brmcoda}
\alias{model.frame.brmcoda}
\title{Extracting the Model Frame from a Formula or Fit from \pkg{brmcoda} object}
\usage{
\method{model.frame}{brmcoda}(formula, ...)
}
\arguments{
\item{formula}{A \code{brmcoda} object.}

\item{...}{Further arguments to be passed to methods.}
}
\description{
Extracting the Model Frame from a Formula or Fit from \pkg{brmcoda} object
}
