% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvcoda.R
\name{mvcoda}
\alias{mvcoda}
\title{A series of functions for single level multivariate compositional models}
\usage{
mvcoda(data, formula, sbp, parts, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table}
containing data of all variables used in the analysis.
It must include a composition and a ID variable. Required.}

\item{formula}{A object of class \code{\link{brmsformula}}, \code{\link{mvbrmsformula}}:
A symbolic description of the model to be fitted.}

\item{sbp}{A signary matrix indicating sequential binary partition. Required.
Details of the model specification can be found in \code{\link{mvbrmsformula}}.}

\item{parts}{A character vector specifying the names of compositional variables. Required.}

\item{...}{Further arguments passed to \code{\link{brm}}.}
}
\value{
A list of results. TODO
}
\description{
Fit Bayesian generalised (non-)linear multivariate compositional model
via full Bayesian inference using Stan,
when composition is the outcome.
}
\examples{
data(mcompd)
data(sbp)
## run mvcoda
mv <- mvcoda(mcompd, mvbind(ilr1, ilr2, ilr3, ilr4) ~ STRESS + Age + Female,
              sbp, c("TST", "WAKE", "MVPA", "LPA", "SB"))

}
