% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{nobs.brmcoda}
\alias{nobs.brmcoda}
\title{Extract Number of Observations from \code{brmcoda} object}
\usage{
\method{nobs}{brmcoda}(object, ...)
}
\arguments{
\item{object}{A \code{brmcoda} object.}

\item{...}{Further arguments to be passed to methods.}
}
\description{
Extract Number of Observations from \code{brmcoda} object
}
