% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.substitution}
\alias{plot.substitution}
\title{Substitution Plot}
\usage{
\method{plot}{substitution}(x, ...)
}
\arguments{
\item{x}{A \code{\link{substitution}} class object.}

\item{...}{Further components to the plot, followed by a plus sign (+).}
}
\value{
A ggplot graph object showing the estimated difference in outcome when
each pair of compositional variables are substituted for a specific time.
}
\description{
Make a plot of \code{\link{substitution}} model results.
}
