% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.substitution}
\alias{plot.substitution}
\title{Substitution Plot}
\usage{
\method{plot}{substitution}(x, to, ref, level, ...)
}
\arguments{
\item{x}{A \code{\link{substitution}} class object.}

\item{to}{A character value or vector specifying the names of the compositional parts
that were reallocated to in the model.}

\item{ref}{A character value of ((\code{"grandmean"} or \code{"clustermean"} or \code{"users"}),}

\item{level}{A character value of (\code{"between"}, \code{"within"}), or \code{"aggregate"}).}

\item{...}{Further components to the plot, followed by a plus sign (+).}
}
\value{
A ggplot graph object showing the estimated difference in outcome when
each pair of compositional variables are substituted for a specific time.
}
\description{
Make a plot of \code{\link{substitution}} model results.
}
