% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot.R
\name{plotsub}
\alias{plotsub}
\title{Generate multilevel substitution plot}
\usage{
plotsub(data, iv, dv, font = "Times New Roman")
}
\arguments{
\item{data}{\code{wsub}, \code{bsub}, \code{wsubmargin}, \code{bsubmargin}.}

\item{iv}{A character string indicating name of the compostional predictor variable}

\item{dv}{A character string indicating the name of the outcome variable.}

\item{font}{A character string indicating the name of preferred font. Default is Times New Roman.}
}
\value{
A ggplot graph object showing the estimated difference in outcome when

each pair of compositional variables are substituted for a specific time.
}
\description{
This function is designed to generate a plot showing the
estimated differences in outcomes when compositional variables
are substituted for a specific period of time
}
\examples{
## TODO

data(mcompd)
plotsub(data = wsubmargintest$SB, iv = "sleep", dv = "stress")

}
