% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoda.R
\name{plotsub}
\alias{plotsub}
\title{Substitution plot}
\usage{
plotsub(data, x, y, ...)
}
\arguments{
\item{data}{A dataset to use for plot.
It must be a component of a list resulted from one of the following functions:
\code{\link{wsub}}, \code{\link{bsub}}, \code{\link{wsubmargins}}, \code{\link{bsubmargins}}.}

\item{x}{A character string specifying name of the compostional predictor variable.}

\item{y}{A character string specifying the name of the outcome variable.}

\item{...}{Further arguments passed to \code{\link{ggplot}}.}
}
\value{
A ggplot graph object showing the estimated difference in outcome when
each pair of compositional variables are substituted for a specific time.
}
\description{
This function is useful for visualising the
estimated differences in outcomes when compositional variables
are substituted for a specific period of time.
}
