% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possub.R
\name{possub}
\alias{possub}
\title{Possible Pairwise Substitution}
\usage{
possub(parts)
}
\arguments{
\item{parts}{A character vector specifying the names of compositional variables to be used.}
}
\value{
A data table of all possible pairwise substitution.
}
\description{
Makes a dataset of all possible pairwise substitution of a composition.
}
\examples{

data(mcompd)

ps1 <- possub(parts = c("TST", "WAKE", "MVPA", "LPA", "SB"))
ps2 <- possub(c("WAKE", "MVPA", "LPA", "SB"))

print(ps2)

## cleanup
rm(mcompd, ps1, ps2)
}
