% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possub.R
\name{possub}
\alias{possub}
\title{Possible Pairwise Substitution}
\usage{
possub(data, composition)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table}
containing the composition used in the analysis. Required.}

\item{composition}{A character vector specifying the names of compositional variables. Required.}
}
\value{
A data table of all possible pairwise substitution.
}
\description{
This function generate a dataset of all possible pairwise substitution of a composition.
}
\examples{

data(mcompd)

ps1 <- possub(data = mcompd, composition = c("TST", "WAKE", "MVPA", "LPA", "SB"))
ps2 <- possub(data = mcompd, composition = c("WAKE", "MVPA", "LPA", "SB"))

print(s2)

## cleanup
rm(mcompd, ps1, ps2)
}
