% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmcoda-methods.R
\name{pp_check.brmcoda}
\alias{pp_check.brmcoda}
\alias{pp_check}
\title{Posterior Predictive Checks for \code{brmcoda} Objects}
\usage{
\method{pp_check}{brmcoda}(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmcoda}.}

\item{...}{Further arguments passed to \code{\link[brms]{predict.brmsfit}}
as well as to the PPC function specified in \code{type}.}
}
\description{
Perform posterior predictive checks with the help of the \pkg{bayesplot} package.
}
\seealso{
\code{\link[brms:pp_check.brmsfit]{pp_check.brmsfit}}
}
