% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.brmcoda}
\alias{print.brmcoda}
\title{Print a summary for a fitted \code{brmsfit} model in a \code{brmcoda} object}
\usage{
\method{print}{brmcoda}(x, ...)
}
\arguments{
\item{x}{An object of class \code{brmcoda}.}

\item{...}{Other arguments passed to \code{summary.brmsfit}}
}
\description{
Print a summary for a fitted \code{brmsfit} model in a \code{brmcoda} object
}
\seealso{
\code{\link{summary.brmcoda}}
}
