% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.compilr}
\alias{print.compilr}
\title{Print a summary for a \code{compilr} object}
\usage{
\method{print}{compilr}(x, ...)
}
\description{
Print a summary for a \code{compilr} object
}
\seealso{
\code{\link{summary.compilr}}
}
