% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-print.R
\name{print.mlsim_data}
\alias{print.mlsim_data}
\title{Print Simulated Data}
\usage{
\method{print}{mlsim_data}(x, ...)
}
\arguments{
\item{x}{An \code{mlsim_data} object returned by \code{\link[=simulate_data]{simulate_data()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a compact overview of an \code{mlsim_data} object.
}
\examples{
sim <- simulate_data(
  n = 3,
  generators = list(
    x = gen_mvn("x", fixed_intercept = 0, residual_cov = 1)
  )
)
print(sim)

}
