% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psub}
\alias{psub}
\title{Possible Pairwise Substitutions}
\format{
A data table containing 5 variables.
\describe{
\item{V1}{first compositional variable}
\item{V2}{second compositional variable}
\item{V3}{third compositional variable}
\item{V4}{fourth compositional variable}
\item{V5}{fifth compositional variable}
}
}
\usage{
psub
}
\description{
A dataset containing possible pairwise subsitutions.
}
\keyword{datasets}
