% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{residuals.brmcoda}
\alias{residuals.brmcoda}
\title{Posterior Draws of Residuals/Predictive Errors}
\usage{
\method{residuals}{brmcoda}(object, ...)
}
\value{
An S x N \code{array} of predictive error draws, where S is the
number of posterior draws and N is the number of observations.
}
\description{
Compute posterior draws of residuals/predictive errors
}
