% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbp}
\alias{sbp}
\title{Sequential Binary Partition}
\format{
A matrix with 5 columns and 4 rows.
\describe{
\item{V1}{first compositional variable}
\item{V2}{second compositional variable}
\item{V3}{third compositional variable}
\item{V4}{fourth compositional variable}
\item{V5}{fifth compositional variable}
}
}
\usage{
sbp
}
\description{
A matrix of sequential binary partition.
}
\keyword{datasets}
