% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbp}
\alias{sbp}
\title{Sequential Binary Partition}
\format{
A matrix with 5 columns and 4 rows.
\describe{
\item{TST}{first compositional variable}
\item{WAKE}{second compositional variable}
\item{MVPA}{third compositional variable}
\item{LPA}{fourth compositional variable}
\item{SB}{fifth compositional variable}
}
}
\usage{
sbp
}
\description{
A matrix of sequential binary partition.
}
\keyword{datasets}
