% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcoda.R
\name{subcoda}
\alias{subcoda}
\title{A series of functions for single level compositional models}
\usage{
subcoda(data, composition, sbp, formula, minute, substitute, ...)
}
\arguments{
\item{data}{A composition or dataset of composition. Required.}

\item{composition}{A string character indicating the names of compositional variables in \code{data}.}

\item{sbp}{A signary matrix indicating sequential binary partition. Required.}

\item{...}{Further arguments passed to \code{\link{brm}}.}
}
\value{
A list with seven elements.
\itemize{
\item{\code{BrmsModel}}{ An object of class \code{brmsfit}, which contains the posterior draws
along with many other useful information about the model.}
\item{\code{SubstitutionResults}}{ A list containing the result of isotemporal substitution model
for each compositional variable.}
\item{\code{ILR}}{ Isometric log ratio transform of composition.}
\item{\code{Comp}}{A vector of class \code{acomp} representing one composition
or a matrix of class \code{acomp} representing multiple closed compositions each in one row.}
\item{\code{data}}{ The user's dataset or imputed dataset if the input data contains zeros.}
\item{\code{sbp}}{ The user-defined sequential binary partition matrix.}
\item{\code{composition}}{ Names of compositional variables.}
}
\description{
The \code{\link{subcoda}} function wraps a series of
The workflows is as follow:
The function firstly compute composition and ILR coordinates,
then fit a Bayesian single-level compositional model to compositional predictors
and users' choices of outcomes and other covariates. Finally, it computes the
isotemporal compositional substitution model according to the user's desired time period.
}
\examples{

data(mcompd)
data(sbp)

davg <- mcompd[, .(SB = mean(SB, na.rm = TRUE),
                   LPA = mean(LPA, na.rm = TRUE),
                   MVPA = mean(MVPA, na.rm = TRUE),
                   TST = mean(TST, na.rm = TRUE),
                   WAKE = mean(WAKE, na.rm = TRUE),
                   TRESS = mean(STRESS, na.rm = TRUE),
                   Age = mean (Age, na.rm = TRUE),
                   Female = na.omit(Female)[1]),
                   by = .(ID)]

scoda <- subcoda(data = davg, composition = c("TST", "WAKE", "MVPA", "LPA", "SB"), 
                       sbp = sbp, minute = 10, substitute = posubtest,
                       formula = STRESS ~ ilr1 + ilr2 + ilr3 + ilr4)
 
 ## cleanup
 rm(mcompd, sbp, davg, scoda)
}
