% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcoda.R
\name{subcoda}
\alias{subcoda}
\title{A series of functions for single level models}
\usage{
subcoda(compilr)
}
\arguments{
\item{x}{A vector.}
}
\value{
The vector coerced to numeric vector in a data table.
}
\description{
This provides a few sentence description about the example function.
}
\examples{

# just use total comp for this

}
