% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.compilr}
\alias{summary.compilr}
\title{Create a summary of a \code{compilr} object}
\usage{
\method{summary}{compilr}(
  object,
  x = c("composition", "logratio"),
  level = c("between", "within", "total"),
  digits = 3,
  ...
)
}
\arguments{
\item{object}{An object of class \code{compilr}.}
}
\description{
Create a summary of a \code{compilr} object
}
