% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.complr}
\alias{summary.complr}
\title{Create a Summary of a \code{complr} object}
\usage{
\method{summary}{complr}(object, weight = c("equal", "proportional"), ...)
}
\arguments{
\item{object}{An object of class \code{complr}.}

\item{weight}{A character value specifying the weight to use in calculation of the reference composition.
If \code{"equal"}, give equal weight to units (e.g., individuals).
If \code{"proportional"}, weights in proportion to the frequencies of units being averaged
(e.g., observations across individuals).
Default is \code{equal}.}

\item{...}{generic argument, not in use.}
}
\description{
Create a Summary of a \code{complr} object
}
\examples{
cilr <- complr(data = mcompd, sbp = sbp, 
               parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), 
               idvar = "ID")
summary(cilr)
}
