% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-print.R
\name{summary.mlsim_data}
\alias{summary.mlsim_data}
\title{Summarize Simulated Data}
\usage{
\method{summary}{mlsim_data}(object, ...)
}
\arguments{
\item{object}{An \code{mlsim_data} object returned by \code{\link[=simulate_data]{simulate_data()}}.}

\item{...}{Ignored.}
}
\value{
A \code{summary.mlsim_data} object with \code{design} and \code{generators} tables.
}
\description{
Build a compact summary of an \code{mlsim_data} object.
}
\examples{
sim <- simulate_data(n = 3, generators = list(x = gen_mvn("x", fixed_intercept = 0, residual_cov = 1)))
summary(sim)

}
