% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.substitution}
\alias{summary.substitution}
\title{Summary \code{\link{substitution}}}
\usage{
\method{summary}{substitution}(object, delta, to, from, ref, level, digits = 2, ...)
}
\arguments{
\item{object}{A \code{\link{substitution}} object containing the output of substitution models.}

\item{delta}{A integer, numeric value or vector indicating the desired \code{delta}
at which substitution results should be summarised.
Default to all \code{delta} available in the \code{\link{substitution}} object.}

\item{to}{A character value or vector specifying the names of the compositional parts
that were reallocated to in the model.}

\item{from}{A character value or vector specifying the names of the compositional parts
that were reallocated from in the model.}

\item{ref}{Either a character value or vector ((\code{grandmean} and/or \code{clustermean} or \code{users}),
Default to all \code{ref} available in the \code{\link{substitution}} object .}

\item{level}{A character string or vector (\code{between} and/or \code{within}).
Default to all \code{level} available in the \code{\link{substitution}} object .}

\item{digits}{A integer value used for number formatting. Default to 3.}

\item{...}{generic argument, not in use.}
}
\value{
A summary of \code{\link{substitution}} object.
\itemize{
\item{\code{Mean}}{ Posterior means.}
\item{\code{CI_low}} and \item{\code{CI_high}}{ 95\% credible intervals.}
\item{\code{Delta}}{ Amount substituted across compositional parts.}
\item{\code{From}}{ Compositional part that is substituted from.}
\item{\code{To}}{ Compositional parts that is substituted to.}
\item{\code{Level}}{ Level where changes in composition takes place. Either }
\item{\code{Reference}}{ Either \code{grandmean}, \code{clustermean}, or \code{users}}
}
}
\description{
This method allows for summarising an existing \code{\link{substitution}} object.
}
