% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complr-methods.R
\name{var.complr}
\alias{var.complr}
\title{Variance of compositions presented in a \code{complr} object.}
\usage{
var.complr(x, weight = c("equal", "proportional"), ...)
}
\arguments{
\item{x}{An object of class \code{complr}.}

\item{weight}{A character value specifying the weight to use in calculation of the reference composition.
If \code{"equal"}, give equal weight to units (e.g., individuals).
If \code{"proportional"}, weights in proportion to the frequencies of units being averaged
(e.g., observations across individuals)
Default is \code{equal}.}

\item{...}{generic argument, not in use.}
}
\description{
Variance of compositions presented in a \code{complr} object.
}
\examples{
cilr <- complr(data = mcompd, sbp = sbp, 
                parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), 
                idvar = "ID")
var.complr(cilr)
}
