% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{VarCorr.brmcoda}
\alias{VarCorr.brmcoda}
\alias{VarCorr}
\title{Extract Variance and Correlation Components}
\usage{
\method{VarCorr}{brmcoda}(object, ...)
}
\value{
A list of lists (one per grouping factor), each with
three elements: a matrix containing the standard deviations,
an array containing the correlation matrix, and an array
containing the covariance matrix with variances on the diagonal.
}
\description{
This function calculates the estimated standard deviations,
correlations and covariances of the group-level terms
in a multilevel model of class \code{brmsfit}.
For linear models, the residual standard deviations,
correlations and covariances are also returned.
}
