% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{VarCorr.brmcoda}
\alias{VarCorr.brmcoda}
\alias{VarCorr}
\title{Extract Variance and Correlation Components}
\usage{
\method{VarCorr}{brmcoda}(x, ...)
}
\arguments{
\item{x}{An object of class \code{brmcoda}.}

\item{...}{Further arguments passed to \code{\link{VarCorr.brmsfit}}}
}
\value{
A list of lists (one per grouping factor), each with
three elements: a matrix containing the standard deviations,
an array containing the correlation matrix, and an array
containing the covariance matrix with variances on the diagonal.
}
\description{
Calculates the estimated standard deviations,
correlations and covariances of the group-level terms
of the \code{brmsfit} object in a \code{brmcoda} object.
}
\seealso{
\code{\link[brms:VarCorr.brmsfit]{VarCorr.brmsfit}}
}
