% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{vcov.brmcoda}
\alias{vcov.brmcoda}
\title{Covariance and Correlation Matrix of Population-Level Effects}
\usage{
\method{vcov}{brmcoda}(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmcoda}.}

\item{...}{Further arguments passed to \code{\link{vcov.brmsfit}}}
}
\value{
covariance or correlation matrix of population-level parameters
}
\description{
Get a point estimate of the covariance or
correlation matrix of population-level parameters
of the \code{brmsfit} object in a \code{brmcoda} object.
}
\seealso{
\code{\link[brms:vcov.brmsfit]{vcov.brmsfit}}
}
