% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{vcov.brmcoda}
\alias{vcov.brmcoda}
\title{Covariance and Correlation Matrix of Population-Level Effects}
\usage{
\method{vcov}{brmcoda}(object, ...)
}
\arguments{
\item{object}{An object of class \code{brmcoda}.}

\item{...}{Further arguments passed to \code{\link{predict.brmsfit}}
that control additional aspects of prediction.}

\item{correlation}{Logical; if \code{FALSE} (the default), compute
the covariance matrix, if \code{TRUE}, compute the correlation matrix.}
}
\value{
covariance or correlation matrix of population-level parameters
}
\description{
Get a point estimate of the covariance or
correlation matrix of population-level parameters
}
