% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsub.R
\name{wsub}
\alias{wsub}
\title{Within-person Substitution (from sample's compositional mean)}
\usage{
wsub(data, substitute, minute = 60)
}
\arguments{
\item{data}{(2) relevant inputs for substitution model.}

\item{substitute}{A data frame or data table indicating the possible substitution of variables. This dataset can be computed using \code{possub}. Required.}

\item{minute}{A integer or numeric value indicating the minute that compositional variable are substituted to/from. Default is 60L.}
}
\value{
\subsection{TODO}{
}
}
\description{
Estimate the difference in outcomes
when compositional variables are substituted for a specific time period
at within-person level.
}
\examples{
## TODO
wsubtest <- wsub(data = mcm, substitute = ps, minute = 10)
}
