% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsub.R
\name{wsub}
\alias{wsub}
\title{Within-person Substitution Model}
\usage{
wsub(object, substitute, minute = 60L, regrid = NULL, summary = TRUE, ...)
}
\arguments{
\item{object}{A \code{\link{brmcoda}} object.}

\item{substitute}{A \code{data.frame} or \code{data.table} of the possible substitution of variables.
This dataset can be computed using function \code{possub}. Required.}

\item{minute}{A integer or numeric value indicating the maximum minute
for which substitution model is desired.
Default to \code{60L} (i.e., the model loops through 1:60L minutes).}

\item{regrid}{If non-\code{NULL}, a \code{data.table} of reference grid consisting
of combinations of covariates over which predictions are made.
Otherwise, the reference grid is constructed via \code{\link{ref_grid}}.}

\item{summary}{A logical value.
Should estimated marginal means at each level of the reference grid (\code{FALSE})
or their average (\code{TRUE}) be returned? Default to \code{TRUE}.}

\item{...}{Additional arguments to be passed to \code{\link{describe_posterior}}.}
}
\value{
A list containing the result of isotemporal multilevel substitution model.
Each elements of the list is the substitution estimation for a compositional variables,
which include at least six elements.
\itemize{
\item{\code{Mean}}{ Posterior means.}
\item{\code{CI_low}} and \item{\code{CI_high}}{ 95\% credible intervals.}
\item{\code{MinSubstituted}}{ Minute substituted within the composition.}
\item{\code{Substitute}}{Compositional variables to be substituted from/to.}
\item{\code{Predictor}}{Central compositional variable to be substituted from/to.}
}
}
\description{
Estimate the difference in outcomes
when compositional variables are substituted for a specific time period
at \code{within-person} level. The resulting \code{wsub} encapsulates
substitution estimation across all compositional variables
present in the \code{\link{brmcoda}} object.

Notes: The reference composition for substitution model
is the compositional mean of the dataset provided.
For average marginal effect, consider using \code{\link{wsubmargins}}.
}
\examples{
\dontrun{
data(mcompd)
data(sbp)
data(psub)
cilr <- compilr(data = mcompd, sbp = sbp, 
                parts = c("TST", "WAKE", "MVPA", "LPA", "SB"), idvar = "ID")

# model with compositional predictor at between and within-person levels
m <- brmcoda(compilr = cilr, 
             formula = STRESS ~ bilr1 + bilr2 + bilr3 + bilr4 + 
                                wilr1 + wilr2 + wilr3 + wilr4 + (1 | ID), 
             chain = 1, iter = 500)
             
subm <- wsub(object = m, substitute = psub, minute = 5)
}
}
