% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsubmargins.R
\name{wsubmargins}
\alias{wsubmargins}
\title{Estimating Average Marginal Effects for Within-person Isotemporal Substitution Model.}
\usage{
wsubmargins(object, substitute, minute = 60, ...)
}
\arguments{
\item{object}{A \code{\link{brmcoda}} object. Required.}

\item{substitute}{A data frame or data table indicating the possible substitution of variables.
This dataset can be computed using \code{possub}. Required.}

\item{minute}{A integer or numeric value indicating the maximum minute for which substitution model is desired.
Default to 60L. In this case, the model loops through 1:60L minutes.}

\item{...}{Additional arguments to be passed to \code{\link{describe_posterior}}.}
}
\value{
A list containing the result of isotemporal multilevel substitution model.
Each elements of the list is the substitution estimation for a compositional variable.
}
\description{
Using a fitted model object, estimates the difference in outcomes
when compositional variables are substituted for a specific period
at within-person level. The resulting \code{bsubmargins} encapsulates
substitution estimation across all compositional variables present
in the \code{\link{brmcoda}} object.
}
\examples{

data(sbp)
data (mcompd)

ps <- possub(count = 5, data = mcompd[, 1:6], idvar = "ID")

## add object - change name to data

wsubmarginstest <- wsubmargins(object = adjbrmcodatest, substitute = ps, minute = 10)
}
