% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsubmargins.R
\name{wsubmargins}
\alias{wsubmargins}
\title{Margianl Effects of Within-person Substitution}
\usage{
wsubmargins(data, substitute, minute = 60)
}
\arguments{
\item{data}{A fitted \code{brms} model object. Required.}

\item{substitute}{This dataset can be computed using \code{possub}. Required.}

\item{minute}{are substituted to/from. Default is 60L.}
}
\value{
A list of outputs from substitution model.

Fitted values extracted from the object object. Mean and intervals.
}
\description{
Estimate the difference in outcomes
when compositional variables are substituted for a specific time period
at within-person level.
}
\examples{

data(sbp)
data (mcompd)

posubtest <- possub(count = 5, data = mcompd[, 1:6], idvar = "ID")

## add object - change name to data

wsubmarginstest <- wsubmargins(data = brmcodatest, substitute = posubtest, minute = 10)
}
